v0.7.5 (#9)
* `r-reticulate` environment setup on install (#8)

Changes the way reticulate installs the python package upon loading/installing the R package. The r-reticulate environment is now manually installed with reticulate::virtualenv_create() and reticulate::virtualenv_install(), in place of reticulate::py_install().

* `pysam` install fix (#10)

Resolves the R client's issues with installing `pysam`  by enforcing:
* `setuptools` version `58.0.0` in the `python` env
* `libpython` libraries within the `python` env used.
 
If necessary, reinstalls `setuptools`, re-creates the `r-reticulate` env, and/or prompts the user to install a proper version of `python` (`3.6` or greater, with `libpython` libraries).

Prevents the `r-reticulate` virtual env from being re-created if it already exists and proper versions of `python` and `setuptools` are detected.

* Install restructuring

Restructures the installation code into `install.R`, which now builds the `r-reticulate` env for use with Toolchest. `.onLoad()` now calls `install_toolchest()`, which checks/builds the env and installs the Python package.